+++ app/email.py
@@ -1,7 +1,5 @@
 from flask_mail import Message
 from app import mail
-from flask import render_template
-from app import app
 
 def send_email(subject, sender, recipients, text_body, html_body):
     msg = Message(subject, sender=sender, recipients=recipients)
@@ -9,12 +7,3 @@ def send_email(subject, sender, recipients, text_body, html_body):
     msg.html = html_body
     mail.send(msg)
 
-def send_password_reset_email(user):
-    token = user.get_reset_password_token()
-    send_email('[Microblog] Reset Your Password',
-               sender=app.config['ADMINS'][0],
-               recipients=[user.email],
-               text_body=render_template('email/reset_password.txt',
-                                         user=user, token=token),
-               html_body=render_template('email/reset_password.html',
-                                         user=user, token=token))
