+++ app/templates/auth/login.html
@@ -9,9 +9,9 @@
         </div>
     </div>
 <br>
-    <p> New User <a href="{{ url_for('register') }}">Click to Register!</a></p>
+    <p> New User <a href="{{ url_for('auth.register') }}">Click to Register!</a></p>
     <p>
         Forgot Your Password?
-        <a href="{{ url_for('reset_password_request') }}">Click to Reset It</a>
+        <a href="{{ url_for('auth.reset_password_request') }}">Click to Reset It</a>
     </p>
 {% endblock %}
