+++ app/templates/email/reset_password.html
@@ -1,7 +1,7 @@
 <p>Dear {{ user.username }},</p>
 <p>
     To reset your password
-    <a href="{{ url_for('reset_password', token=token, _external=True) }}">
+    <a href="{{ url_for('auth.reset_password', token=token, _external=True) }}">
         click here
     </a>.
 </p>
