+++ app/templates/user.html
@@ -8,11 +8,11 @@
             {% if user.about_me %}<p>{{ user.about_me }}</p>{% endif %}
             {% if user.last_seen %}<p>Last seen on: {{ user.last_seen }}</p>{% endif %}
             {% if user == current_user %}
-                <p><a href="{{ url_for('edit_profile') }}">Edit your profile</a></p>
+                <p><a href="{{ url_for('main.edit_profile') }}">Edit your profile</a></p>
             {% elif not current_user.is_following(user) %}
-                <p><a href="{{ url_for('follow', username=user.username) }}">Follow</a></p>
+                <p><a href="{{ url_for('main.follow', username=user.username) }}">Follow</a></p>
             {% else %}
-                <p><a href="{{ url_for('unfollow', username=user.username) }}">Unfollow</a></p>
+                <p><a href="{{ url_for('main.unfollow', username=user.username) }}">Unfollow</a></p>
             {% endif %}
             </td>
         </tr>
